Add files via upload
I noticed when I was done that I had received feedback on the previous portfolio saying that one could get Visual Code Studio to upload all the changes you do to GitHub after each session. It is my own fault I did not know how to do this, and unfortunately it is the same case with this portfolio submission. I will be asking about it in class so that I can make it right for the next one. I have been saving locally after every session, and Visual Code Studio has allowed me to pick up straight from where I left off, but next time I want to change that.